+++ packages/wrapper/CHANGELOG.md
@@ -2,9 +2,20 @@
 
 All notable changes to npm module [monaco-editor-wrapper](https://www.npmjs.com/package/monaco-editor-wrapper) are documented in this file.
 
+## [6.2.4] - 2025-02-06
+
+- Fix non dynamic import of @codingame/monaco-vscode-views-service-override [#844](https://github.com/TypeFox/monaco-languageclient/pull/844)
+- Updated all `@codingame/monaco-vscode` packages to `13.1.4`.
+- Updated to `monaco-languageclient@9.2.4`.
+
+## [6.2.3] - 2025-02-04
+
+- Updated all `@codingame/monaco-vscode` packages to `13.1.3`
+- Updated to `monaco-languageclient@9.2.3`.
+
 ## [6.2.2] - 2025-02-03
 
-- Updated all `@codingame/monaco-vscode` packages to `13.1.1`
+- Updated all `@codingame/monaco-vscode` packages to `13.1.2`
 - Updated to `monaco-languageclient@9.2.2`.
 
 ## [6.2.1] - 2025-01-31